requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
alembic==1.15.2
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.1.0
pydantic[email]==2.11.3
pydantic-settings==2.9.1
pytest==8.4.0
httpx==0.28.1
